Refactor - moved files into rough subfolder categories
