+++ src/BoardSetup/Board.py
@@ -1,12 +1,12 @@
-from src.Bishop import Bishop
-from src.Coordinate import Coordinate as C
-from src.King import King
-from src.Knight import Knight
-from src.Move import Move
-from src.Pawn import Pawn
-from src.Queen import Queen
-from src.Rook import Rook
-from termcolor import colored
+from src.BoardSetup.Coordinate import Coordinate as C
+from src.BoardSetup.Move import Move
+from src.BoardSetup.termcolor import colored
+from src.ChessPieces.Bishop import Bishop
+from src.ChessPieces.King import King
+from src.ChessPieces.Knight import Knight
+from src.ChessPieces.Pawn import Pawn
+from src.ChessPieces.Queen import Queen
+from src.ChessPieces.Rook import Rook
 
 WHITE = True
 BLACK = False
