+++ src/BoardSetup/termcolor.py
@@ -24,8 +24,8 @@
 """ANSII Color formatting for output in terminal."""
 
 from __future__ import print_function
-import os
 
+import os
 
 __ALL__ = ['colored', 'cprint']
