+++ src/ChessPieces/King.py
@@ -1,6 +1,6 @@
-from src.Coordinate import Coordinate as C
-from src.Move import Move
-from src.Piece import Piece
+from src.BoardSetup.Coordinate import Coordinate as C
+from src.BoardSetup.Move import Move
+from src.ChessPieces.Piece import Piece
 
 WHITE = True
 BLACK = False
